+++ listenbrainz/webserver/static/js/src/UserPageHeading.test.tsx
@@ -22,8 +22,8 @@ import * as React from "react";
 import { mount, shallow } from "enzyme";
 import UserPageHeading from "./UserPageHeading";
 import FollowButton from "./FollowButton";
-import ReportUserButton from "./ReportUser";
-import ReportUserModal from "./ReportUserModal";
+import ReportUserButton from "./report-user/ReportUser";
+import ReportUserModal from "./report-user/ReportUserModal";
 import GlobalAppContext, { GlobalAppContextT } from "./GlobalAppContext";
 import APIService from "./APIService";
