+++ listenbrainz/webserver/static/js/tests/missing-mb-data/MissingMBData.test.tsx
@@ -1,16 +1,16 @@
 import * as React from "react";
 import { mount } from "enzyme";
 
-import * as missingDataProps from "../tests/__mocks__/missingMBDataProps.json";
+import * as missingDataProps from "../__mocks__/missingMBDataProps.json";
 import {
   youtube,
   spotify,
   user,
-} from "../tests/__mocks__/missingMBDataProps.json";
+} from "../__mocks__/missingMBDataProps.json";
 
-import MissingMBDataPage from "./MissingMBData";
-import GlobalAppContext, { GlobalAppContextT } from "./GlobalAppContext";
-import APIService from "./APIService";
+import MissingMBDataPage from "../../src/missing-mb-data/MissingMBData";
+import GlobalAppContext, { GlobalAppContextT } from "../../src/GlobalAppContext";
+import APIService from "../../src/APIService";
 
 // Font Awesome generates a random hash ID for each icon everytime.
 // // Mocking Math.random() fixes this
